+++ service/list_urls.py
@@ -1,0 +1,4 @@
+from typing import Dict
+
+async def list_urls() -> Dict[str, str]:
+	return {}
