+++ service/redirect.py
@@ -1,0 +1,2 @@
+async def redirect(short_url: str) -> str:
+	return ''
